Add files via upload
current version has no database connect, player movement, cars have 3x rows of car arrays with minimum of 4 moving in alternating speeds and directions per row, the logs are implemented but currently no functionality other than movement and visually 3 rows of arrays, score is displayed and calculated as needed 50 per win, -50 per loss. Must add Database connection and save data + user's name. Need to implement death logic for river on logs and implement playerRidingLog function to move player along the speed of the log row is it currently on, also muist adjust player to be in line with rows (player is 50 x 50) rows are (50 in width) must have player be inline so we can jump on the log without any clipping issues.